+++ src/shef/mt/features/impl/bb/Feature1115.java
@@ -33,10 +33,10 @@ public class Feature1115 extends Feature {
      */
     @Override
     public void run(Sentence source, Sentence target) {
-        float sourcePP = (Integer) source.getValue("CONJP");
-        float phrasesSource = (Integer) source.getValue("phrase_tags");
-        float targetPP = (Integer) target.getValue("CONJP");
-        float phrasesTarget = (Integer) source.getValue("phrase_tags");
+        float sourcePP = (Float) source.getValue("CONJP");
+        float phrasesSource = (Float) source.getValue("phrase_tags");
+        float targetPP = (Float) target.getValue("CONJP");
+        float phrasesTarget = (Float) source.getValue("phrase_tags");
 
         float sourceNorm = phrasesSource == 0 ? 0 : sourcePP / phrasesSource;
         float targetNorm = phrasesTarget == 0 ? 0 : targetPP / phrasesTarget;
